Moved the help function to mayaUsdUtils.py.
Also moved the helpTable to maya-Usd/resources folder.
