+++ lib/mayaUsd/resources/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(ae)
+add_subdirectory(helpTable)
 add_subdirectory(icons)
 add_subdirectory(scripts)
