+++ lib/mayaUsd/resources/helpTable/helpTableMayaUSD
