+++ lib/mayaUsd/resources/scripts/mayaUsdUtils.py
@@ -193,4 +193,39 @@ def saveWantPayloadLoaded(want):
     opVarName = "mayaUsd_WantPayloadLoaded"
     cmds.optionVar(iv=(opVarName, want))
 
+def showHelpMayaUSD(contentId):
+    """
+    Helper method to display help content.
+
+    Note that this is a wrapper around Maya's showHelp() method and showHelpMayaUSD()
+    should be used for all help contents in Maya USD.
+
+    Example usage of this method:
     
+    - In Python scripts:
+    from mayaUsdUtils import showHelpMayaUSD
+    showHelpMayaUSD("someContentId");
+
+    - In MEL scripts:
+    python(\"from mayaUsdUtils import showHelpMayaUSD; showHelpMayaUSD('someContentId');\")
+    
+    - In C++:
+    MGlobal::executePythonCommand(
+    "from mayaUsdUtils import showHelpMayaUSD; showHelpMayaUSD(\"someContentId\");");
+
+    Input contentId refers to the contentId that is registered in helpTableMayaUSD
+    file which is used to open help pages.
+    """
+    import os
+    
+    try:
+        # Finding the path to helpTableMayaUSD file.
+        dirName = cmds.pluginInfo('mayaUsdPlugin', q=True, p=True)
+        dirName = os.path.dirname(dirName) + "/../../../lib/helpTable/" + "helpTableMayaUSD"
+        # Setting the default helpTable to helpTableMayaUSD
+        cmds.showHelp(dirName, helpTable=True)
+        # Showing the help content
+        cmds.showHelp(contentId)
+    finally:
+        # Restoring Maya's default helpTable
+        cmds.showHelp('helpTable', helpTable=True)
