+++ plugin/adsk/CMakeLists.txt
@@ -2,7 +2,6 @@ message("========== ADSK USD Plugin ==========")
 
 set(INSTALL_DIR_SUFFIX plugin/adsk)
 
-add_subdirectory(helpTable)
 add_subdirectory(icons)
 add_subdirectory(plugin)
 add_subdirectory(scripts)
